for submission - js activty

+++ js/index.js
@@ -13,35 +13,39 @@ btn.addEventListener('change', function(e) {
 
 // post Class
 class Post {
-    constructor (name, pos, text) {
+    constructor (name, pos, text, date, key) {
         this.name = name;
         this.pos = pos;
         this.text = text;
+        this.date = date;
+        this.key = key;
     }
 }
 
 // UI Class
 class UI {
     static displayPosts() {
+        document.querySelector('#post-list').innerHTML = '';  
         const posts = Store.getPosts();
 
         posts.forEach(post => UI.addPostToList(post));
     }
 
     static addPostToList(post) {
-        let currDate = new Date().toLocaleString();
+        
 
         const list = document.querySelector('#post-list');
         const article = document.createElement('article');
         article.innerHTML = `
         <h3>${post.name}</h3>
-        <p><em>${post.pos}</em></p>
+        <p>${post.pos}</p>
         <p>${post.text}</p>
         <section class="mod">
             <i class="edit-btn fa fa-pencil-square-o" title="Edit"></i>
             <i class="delete-btn fa fa-times-circle" title="Delete"></i>
+            <span class="key">${post.key}</span>
         </section>
-        <p class="date">${currDate}</p>
+        <p class="date">${post.date}</p>
         `;
 
         list.appendChild(article);
@@ -54,21 +58,92 @@ class UI {
     }
 
     static editPost(el) {
-        const name = document.querySelector('#name').value
-        const pos = document.querySelector('#position').value
-        const text = document.querySelector('#text').value
-
-        if(el.classList.contains('edit-btn')) {
-            if(name === '' && pos === '' && text === ''){
-                document.querySelector('#name').value = `${el.parentElement.parentElement.firstElementChild.textContent}`;
-                document.querySelector('#position').value = `${el.parentElement.parentElement.firstElementChild.nextElementSibling.textContent}`;
-                document.querySelector('#text').value = `${el.parentElement.parentElement.firstElementChild.nextElementSibling.nextElementSibling.textContent}`;
-                // remove
-                el.parentElement.parentElement.remove();
-            } else {
-                alert('Please clear the fields.');
-            }   
+      if(el.classList.contains('edit-btn')) {
+        document.querySelectorAll('.mod').forEach(element => {
+          element.style.display = 'none';
+        });
+        document.getElementById('name').disabled = true;
+        document.getElementById('position').disabled = true;
+        document.getElementById('text').disabled = true;
+        document.getElementById('post').disabled = true;
+  
+        let name = el.parentElement.parentElement.firstElementChild.textContent;
+        let pos = el.parentElement.parentElement.firstElementChild.nextElementSibling.textContent;
+        let text = el.parentElement.parentElement.firstElementChild.nextElementSibling.nextElementSibling.textContent;
+        let date = el.parentElement.parentElement.lastElementChild.textContent;
+        let key = el.nextElementSibling.nextElementSibling.textContent;
+        
+        el.parentElement.parentElement.firstElementChild.innerHTML = `<input class="feedback-input edit" type="text" value="${name}">`;
+        el.parentElement.parentElement.firstElementChild.nextElementSibling.innerHTML = `<input class="feedback-input edit" type="text" value="${pos}">`;
+        el.parentElement.parentElement.firstElementChild.nextElementSibling.nextElementSibling.innerHTML = `<textarea class="feedback-input">${text}</textarea>`;
+        let save = document.createElement('input');
+        let discard = document.createElement('input');
+        save.setAttribute('id', 'save');
+        save.setAttribute('class', 'post-btn save-btn');
+        save.setAttribute('type', 'submit');
+        save.setAttribute('value', 'Save');
+        discard.setAttribute('id', 'discard');
+        discard.setAttribute('class', 'post-btn discard-btn');
+        discard.setAttribute('type', 'submit');
+        discard.setAttribute('value', 'Discard');
+        
+        el.parentElement.parentElement.appendChild(save);
+        el.parentElement.parentElement.appendChild(discard);
+        let temp = new Post(name, pos, text, date, key)
+        Store.addTemp(temp)
+        // Store.removePost(name, pos, text);
+    }
+  }
+  static savePost(el) {
+      if(el.classList.contains('save-btn')) { 
+        document.querySelectorAll('.mod').forEach(element => {
+          element.style.display = '';
+        });
+        let date = new Date().toLocaleString();
+        date = '(Edited) ' + date;
+        let name = el.parentElement.firstElementChild.firstElementChild.value;
+        let pos = el.parentElement.firstElementChild.nextElementSibling.firstElementChild.value;
+        let text = el.parentElement.firstElementChild.nextElementSibling.nextElementSibling.firstElementChild.value;
+        let key = el.previousElementSibling.previousElementSibling.lastElementChild.textContent;
+        if(name === '' || pos ===  '' || text === '') {
+          alert('Please fill-in all fields.');
+        } else {
+          Store.removePost(key)
+          Store.removeTemp(key)
+          const updatedPost = new Post (name, pos, text, date, null);
+          Store.addPost(updatedPost)
+          UI.displayPosts();
+          document.getElementById('name').disabled = false;
+          document.getElementById('position').disabled = false;
+          document.getElementById('text').disabled = false;
+          document.getElementById('post').disabled = false;
         }
+    }
+  }
+    static discardPost(el) {
+      if(el.classList.contains('discard-btn')) {
+        document.querySelectorAll('.mod').forEach(element => {
+          element.style.display = '';
+        });
+        let key = el.previousElementSibling.previousElementSibling.previousElementSibling.lastElementChild.textContent;
+        let temps = Store.getTemps();
+        temps.forEach((temp) => {
+          if (temp.key == key) {
+              el.parentElement.firstElementChild.innerHTML = temp.name;
+              el.parentElement.firstElementChild.nextElementSibling.innerHTML = temp.pos;
+              el.parentElement.firstElementChild.nextElementSibling.nextElementSibling.innerHTML = temp.text;
+              el.previousElementSibling.previousElementSibling.previousElementSibling.style.display = '';
+              el.previousElementSibling.remove();
+              el.remove();
+              Store.removeTemp(key);
+              document.getElementById('name').disabled = false;
+              document.getElementById('position').disabled = false;
+              document.getElementById('text').disabled = false;
+              document.getElementById('post').disabled = false;
+          }
+        });
+      }
+      
     }
 
     static clearFields() {
@@ -92,21 +167,53 @@ class Store {
     static addPost(post) {
         const posts = Store.getPosts();
         posts.push(post);
+        posts.forEach((post, index) => {
+          post.key = index;
+        });
         localStorage.setItem('posts', JSON.stringify(posts));
     }
-    static removePost(name, text) {
+    static removePost(key) {
         const posts = Store.getPosts();
         posts.forEach((post, index) => {
-            if(post.name === name && post.text === text) {
+            if(post.key == key) {
                 posts.splice(index, 1);
             }
         });
         localStorage.setItem('posts', JSON.stringify(posts));
     }
+
+    static getTemps() {
+      let temps;
+      if(localStorage.getItem('temps') === null) {
+          temps = [];
+      } else {
+          temps = JSON.parse(localStorage.getItem('temps'));
+      }
+      return temps;
+  }
+    static addTemp(temp) {
+      const temps = Store.getTemps();
+      temps.push(temp);
+      localStorage.setItem('temps', JSON.stringify(temps));
+    }
+    static removeTemp(key) {
+      const temps = Store.getTemps();
+      temps.forEach((temp, index) => {
+          if(temp.key == key) {
+              temps.splice(index, 1);
+          }
+      });
+      localStorage.setItem('temps', JSON.stringify(temps));
+  }
+    static resetTemp() {
+      let temps = [];
+      localStorage.setItem('temps', JSON.stringify(temps));
+    }
 }
 
 // Event: display posts
 document.addEventListener('DOMContentLoaded', UI.displayPosts);
+document.addEventListener('DOMContentLoaded', Store.resetTemp);
 
 // Event: add a post
 document.querySelector('#post-form').addEventListener('submit', (e) => {
@@ -115,19 +222,22 @@ document.querySelector('#post-form').addEventListener('submit', (e) => {
     const name = document.querySelector('#name').value;
     const pos = document.querySelector('#position').value;
     const text = document.querySelector('#text').value;
+    const date = new Date().toLocaleString();
     
     // validation
     if(name === '' || pos ===  '' || text === '') {
         alert('Please fill-in all fields.');
     } else {
         // instantiate post
-        const post = new Post (name, pos, text);
+        const post = new Post (name, pos, text, date, null);
     
-        // add post to UI
-        UI.addPostToList(post);
-
         // add post to localStorage
         Store.addPost(post);
+
+        // add post to UI
+        // UI.addPostToList(post);
+
+        UI.displayPosts();
 
         // clear fields
         UI.clearFields();
@@ -136,26 +246,18 @@ document.querySelector('#post-form').addEventListener('submit', (e) => {
 
 // Event: remove or edit a post
 document.querySelector('#post-list').addEventListener('click', (e) => {
-    // remove book from UI
-    UI.deletePost(e.target)
-
+    // remove post from UI
+    if (e.target.classList.contains('delete-btn')) {
+      UI.deletePost(e.target)
+      Store.removePost(e.target.nextElementSibling.textContent);
+    }
     // edit a post
-    UI.editPost(e.target)
-
-    // remove book from localStorage
-    Store.removePost( e.target.parentElement.parentElement.firstElementChild.textContent, e.target.parentElement.previousElementSibling.textContent);
+    UI.editPost(e.target);
+    
+    // save edit 
+    UI.savePost(e.target);
+    
+    UI.discardPost(e.target);
 });
 
-// const initPosts = [
-//     {
-//         name: 'Sir Melvin',
-//         pos: 'Instructor',
-//         text: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Mollitia incidunt nisi tempora inventore consequuntur nam veritatis similique facere? Dolorem, mollitia.'
-//     },
-//     {
-//         name: "Ma'am Annie",
-//         pos: 'Dept. Chair',
-//         text: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Nihil aspernatur quos necessitatibus culpa laboriosam tenetur? '
-//     }
-// ]
-// initPosts.forEach(post => UI.addPostToList(post));
+
